Create Program.cs
Hello, my name is Yizhuo Hu, and I am working on the final project of the object oriented programming class.
Connect 4 is a board game where players take turns placing tiles on a grid, with the goal of forming a line of four tiles to win.
The process of the game is that the first player chooses the column he wants to fall into, and then the second player makes a choice, until one player's pieces are connected into 4, and finally the game ends and the program stops.